requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-SocketIO==4.3.1
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
python-engineio==3.13.1
python-socketio==4.6.0
requests==2.24.0
six==1.15.0
typing-extensions==3.7.4.2
urllib3==1.25.10
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
yarl==1.5.1
